Accept long as a spelling of the integer type
MicroC's integer type could only be written long long. It can now also be
written long, naming the same type: the same 53-bit range, the same
arithmetic, the same lowering. No type kind is added and nothing below the
front end changes.

long long stays what a diagnostic quotes, so an object declared long is still
named long long in a type error. It is the spelling C guarantees at least 64
bits of on every implementation, which long is not: under LLP64 a long is 32
bits, so the strict-subset property is local to the spelling.

spelledPast assumed long long was the only multi-word type, which stopped
holding once long alone named one. It now reports a size word only where the
prefix in front of it already wrote that word, so long double and its
neighbours name the whole spelling rather than failing as a parse error.

The fixture corpus writes long, except where a program's own literal exceeds
2^31. Two do.

Drop the two path gates and the script behind them. They asserted that a file
spelled a path the way the Taskfile said it did, which is a claim about text
rather than about behaviour.

+++ internal/corpus/programs/airlock.c
@@ -18,13 +18,13 @@ const dev pump = d2;
 const dev lowDial = d3;
 const dev highDial = d4;
 
-constexpr long long kIdle = 0;
-constexpr long long kPumping = 1;
-constexpr long long kOpening = 2;
-constexpr long long kFault = 3;
+constexpr long kIdle = 0;
+constexpr long kPumping = 1;
+constexpr long kOpening = 2;
+constexpr long kFault = 3;
 
-long long state;
-long long retries;
+long state;
+long retries;
 double evacuateBelow;
 
 void swap(double *a, double *b) {
@@ -33,7 +33,7 @@ void swap(double *a, double *b) {
     *b = t;
 }
 
-long long gcd(long long a, long long b) {
+long gcd(long a, long b) {
     if (b == 0) {
         return a;
     }
