+++ internal/corpus/programs/bits.c
@@ -4,36 +4,36 @@
 // point unpacks a packed setting into bytes each tick and republishes what it
 // found.
 //
-// Everything here is written as a long long, because everything here is
-// counting rather than measuring: the one fractional value in sight is the raw
-// device reading, which is narrowed with a cast where it is read.
+// Everything here is written as a long, because everything here is counting
+// rather than measuring: the one fractional value in sight is the raw device
+// reading, which is narrowed with a cast where it is read.
 //
-// A long long is still a double in the register -- the chip has no integer
-// registers and no integer type -- so what the C type buys is exactness to 2^53
-// and the operators. The bitwise and shift operators take a long long and
-// nothing else, and everything masked or shifted below stays well inside a
-// byte, so none of it comes near either end of that window.
+// A long is still a double in the register -- the chip has no integer registers
+// and no integer type -- so what the C type buys is exactness to 2^53 and the
+// operators. The bitwise and shift operators take a long and nothing else, and
+// everything they touch below is a byte field or a small running total, so none
+// of it comes near either end of that window.
 
 const dev source = d0;
 const dev report = d1;
 
-constexpr long long kByteMask = 0xff;
-constexpr long long kHighBit = 0x80;
-constexpr long long kDelimiter = ',';
-constexpr long long kDecades[4] = {1, 10, 100, 1000};
+constexpr long kByteMask = 0xff;
+constexpr long kHighBit = 0x80;
+constexpr long kDelimiter = ',';
+constexpr long kDecades[4] = {1, 10, 100, 1000};
 
-long long frame[4];
+long frame[4];
 
-bool isDigit(long long c) {
+bool isDigit(long c) {
     return c >= '0' && c <= '9';
 }
 
-long long digitValue(long long c) {
+long digitValue(long c) {
     return isDigit(c) ? c - '0' : -1;
 }
 
-long long popcount(long long v) {
-    long long n = 0;
+long popcount(long v) {
+    long n = 0;
     do {
         n += v & 1;
         v >>= 1;
@@ -41,11 +41,11 @@ long long popcount(long long v) {
     return n;
 }
 
-long long checksum(long long buf[], long long len) {
-    long long total = 0;
-    long long i = 0;
+long checksum(long buf[], long len) {
+    long total = 0;
+    long i = 0;
     while (i < len) {
-        long long b = buf[i] & kByteMask;
+        long b = buf[i] & kByteMask;
         i++;
         if (b == 0) {
             continue;
@@ -61,19 +61,19 @@ long long checksum(long long buf[], long long len) {
     return total & kByteMask;
 }
 
-long long scaled(long long v, long long decade) {
+long scaled(long v, long decade) {
     return decade < 4 ? v * kDecades[decade] : v;
 }
 
 void announce(void) {
-    long long lights = __ic_hash("StructureWallLight");
-    long long consoles = __ic_hash("StructureConsoleLED5");
+    long lights = __ic_hash("StructureWallLight");
+    long consoles = __ic_hash("StructureConsoleLED5");
     __ic_store_batch(lights, On, 1);
     __ic_store_batch_named(consoles, __ic_hash("Airlock"), Setting, popcount(kByteMask));
 }
 
-void unpack(long long packed) {
-    for (long long i = 0; i < 4; i++) {
+void unpack(long packed) {
+    for (long i = 0; i < 4; i++) {
         frame[i] = (packed >> (i * 8)) & kByteMask;
     }
 }
@@ -82,10 +82,10 @@ void main(void) {
     announce();
 
     while (true) {
-        unpack((long long)__ic_load(source, Setting));
+        unpack((long)__ic_load(source, Setting));
 
-        long long digit = digitValue(frame[0]);
-        long long value = 0;
+        long digit = digitValue(frame[0]);
+        long value = 0;
         if (digit >= 0) {
             value = scaled(digit, popcount(frame[1]) & 3);
         }
