+++ internal/corpus/programs/liquid_train.c
@@ -20,14 +20,14 @@ const dev drain = d1;
 
 constexpr double kContaminated = 0.02;
 constexpr double kSingleFault = 0.008;
-constexpr long long kModePercent = 1;
+constexpr long kModePercent = 1;
 
 double held;
 
 // contamination is everything at one tap that is not water, as a fraction of
 // what the pipe holds. Polluted water is counted with the rest: it leaves the
 // train on the same side the salts do, and a tap running clean has none.
-double contamination(long long analysers, long long tap) {
+double contamination(long analysers, long tap) {
     double total = 0.0;
     total = total + __ic_load_batch_named(analysers, tap, RatioLiquidHydrochloricAcid, Average);
     total = total + __ic_load_batch_named(analysers, tap, RatioLiquidSodiumChloride, Average);
@@ -46,7 +46,7 @@ double contamination(long long analysers, long long tap) {
 // heaviest is the largest single contaminant at one tap, taken over the
 // analysers rather than averaged across them: one pipe of the stage running
 // dirty is the thing a stage average hides.
-double heaviest(long long analysers, long long tap) {
+double heaviest(long analysers, long tap) {
     double worst = 0.0;
     worst = __ic_max(worst, __ic_load_batch_named(analysers, tap, RatioLiquidHydrochloricAcid, BatchMode_Maximum));
     worst = __ic_max(worst, __ic_load_batch_named(analysers, tap, RatioLiquidSodiumChloride, BatchMode_Maximum));
@@ -59,10 +59,10 @@ double heaviest(long long analysers, long long tap) {
 }
 
 void main(void) {
-    long long analysers = __ic_hash("StructureLiquidPipeAnalyzer");
-    long long feed = __ic_hash("Feed");
-    long long middle = __ic_hash("Middle");
-    long long tail = __ic_hash("Tail");
+    long analysers = __ic_hash("StructureLiquidPipeAnalyzer");
+    long feed = __ic_hash("Feed");
+    long middle = __ic_hash("Middle");
+    long tail = __ic_hash("Tail");
 
     while (true) {
         double atFeed = contamination(analysers, feed);
