+++ internal/corpus/programs/thermostat.c
@@ -1,9 +1,9 @@
 // A thermostat holding a room at a setpoint with hysteresis, so the heater
 // does not chatter around the boundary. A room temperature is fractional --
 // 293.15 K is where a habitat starts -- so the reading and the band are
-// doubles, and only the value published back is narrowed to a long long. The
+// doubles, and only the value published back is narrowed to a long. The
 // narrowing is written as a cast because nothing narrows implicitly here: a
-// double reaches a long long only where the line says so.
+// double reaches a long only where the line says so.
 //
 // Both pins say what they are wired to. Nothing enforces that at run time -- a
 // housing reaches whatever the player plugged into it -- but it is what lets the
@@ -18,7 +18,7 @@ constexpr double kHysteresis = 2.0;
 constexpr double kMaxKelvin = 400.0;
 
 bool heaterOn;
-long long lastReading;
+long lastReading;
 
 double clampReading(double v, double lo, double hi);
 
@@ -44,7 +44,7 @@ double clampReading(double v, double lo, double hi) {
 void main(void) {
     while (true) {
         double temp = clampReading(readTemperature(), 0.0, kMaxKelvin);
-        lastReading = (long long)temp;
+        lastReading = (long)temp;
 
         if (temp < kSetpointKelvin - kHysteresis) {
             driveHeater(true);
